third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFESpotLightElement-dom-limitingConeAngle-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'limitingConeAngle' attribute of the SVGFESpotLightElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS spotLightElement.getAttribute('limitingConeAngle') is "45"
PASS spotLightElement.getAttribute('limitingConeAngle') is "15"
PASS successfullyParsed is true

TEST COMPLETE
